fix: close production readiness gaps
schema.prisma: new AccountType enums (HOUSE_RESERVE, TOURNAMENT_ESCROW, PENDING_WITHDRAWAL), remove UNCALLED_BET
config: extensive env vars for all operational thresholds
seed: house escrow accounts idempotent seeding
postgres: migration 002 for house system accounts

user/index.ts: BigInt precision for monetary math, timestamp validation with TTL, PENDING_WITHDRAWAL hold semantics
withdrawal-bot.ts: daily limit tracking via ledger aggregation, broadcast-completion double-entry
deposit-monitor.ts: per-chain lastScannedBlock, deposit double-entry prevention (unique constraint, pre-check, status-guard, reorg detection)
tournament-blinds.ts: NODE_ENV-aware lock TTL ternary
tournaments/index.ts: escrow double-entry prevention (idempotency for buy-in, status-guard for payout)
docker-compose.prod.yml: admin entrypoint override, healthchecks on all services, Redis requirepass auth

utils/tournaments.ts: 20-level geometric blind structure, MAX_TOURNAMENT_TABLES configurable
ws/index.ts + socket-manager.ts: Redis-backed connection tracking and pub/sub, config-driven limits
engine: remove redundant getBlindPositions call, replace magic numbers with named defaults
workers: Redlock consistency across settle-hand/next-hand/timeout, config env vars for all tunables
game-manager.ts: replace legacy Queue shim with typed JobQueues, auto-deal delay configurable
blockchain-service.ts: remove getDerivationPath caching, query directly
sweeper-service.ts: rename MIN_SWEEP_VALUE_USD -> MIN_SWEEP_VALUE_RAW_UNITS
sdk/auth.ts: chainId config, SIWE dedup via nonce atomic consumption
auth/index.ts: NONCE_TTL configurable, NODE_ENV from validated config
admin/README.md: document polling-based withdrawal bot, rename MIN_SWEEP_VALUE_RAW_UNITS
admin/index.ts: shutdown bot.stop() on SIGTERM/SIGINT
docker-compose.prod.yml: remove ALLOWED_SIWE_CHAIN_IDS from admin (unused)

tests: financial integrity, withdrawal, deposit, tournament tests for new ledger entries; new withdrawal-cycle E2E test
typecheck + lint clean, all 1,278 tests passing across 78 suites

+++ packages/api/prisma/schema.sql
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS "Account" (
     "userId" TEXT NOT NULL,
     "currency" TEXT NOT NULL DEFAULT 'USDC',
     "type" TEXT NOT NULL,
-    "balance" INTEGER NOT NULL DEFAULT 0,
+    "balance" BIGINT NOT NULL DEFAULT 0,
     CONSTRAINT "Account_userId_fkey" FOREIGN KEY ("userId") REFERENCES "User" ("id") ON DELETE CASCADE ON UPDATE CASCADE
 );
 
@@ -91,7 +91,7 @@ CREATE TABLE IF NOT EXISTS "IdempotencyRecord" (
 CREATE TABLE IF NOT EXISTS "LedgerEntry" (
     "id" TEXT NOT NULL PRIMARY KEY,
     "accountId" TEXT NOT NULL,
-    "amount" INTEGER NOT NULL,
+    "amount" BIGINT NOT NULL,
     "type" TEXT NOT NULL,
     "referenceId" TEXT,
     "metadata" JSONB,
@@ -110,7 +110,7 @@ CREATE TABLE IF NOT EXISTS "PaymentTransaction" (
     "blockNumber" TEXT,
     "blockHash" TEXT,
     "amountRaw" TEXT NOT NULL,
-    "amountCredit" INTEGER NOT NULL,
+    "amountCredit" BIGINT NOT NULL,
     "fee" TEXT,
     "status" TEXT NOT NULL DEFAULT 'PENDING',
     "recoveryState" TEXT,
